+++ internal_auth/src/epilot/sdkconfiguration.py
@@ -19,9 +19,9 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '0.0.1'
-    sdk_version: str = '1.48.0'
-    gen_version: str = '2.150.0'
-    user_agent: str = 'speakeasy-sdk/python 1.48.0 2.150.0 0.0.1 epilot-internal-auth'
+    sdk_version: str = '1.48.1'
+    gen_version: str = '2.154.1'
+    user_agent: str = 'speakeasy-sdk/python 1.48.1 2.154.1 0.0.1 epilot-internal-auth'
     retry_config: RetryConfig = None
 
     def get_server_details(self) -> tuple[str, dict[str, str]]: